Disclamer
This Script will not be run if you downloaded it as it is not signed generally. You can copy the script to your PowerShell ISE and sign by yourself in order to run it. The original script does not contain any harmful codes and any modified version of the script will not responsible me. To run this sucessfully you will need an internet explorer.